feat: inline column states into collection
