+++ src/DynamicTreeDataGrid/Columns/DynamicColumnList.cs
@@ -41,6 +41,15 @@ public class DynamicColumnList<TModel> : DynamicColumnListBase<TModel>, IDynamic
         base.InsertItem(newIndex, item);
     }
 
+    public IEnumerable<ColumnState> GetColumnStates() {
+        IList<ColumnState> states = [];
+        for (var i = 0; i < Count; i++) {
+            var column = this[i];
+            states.Add(new ColumnState(column.Name) { Visible = column.Visible, Index = i, });
+        }
+        return states;
+    }
+
     public bool ApplyColumnStates(IEnumerable<ColumnState> states) {
 	    try
 	    {
