+++ src/DynamicTreeDataGrid/DynamicHierarchicalTreeDataGridSource.cs
@@ -9,7 +9,6 @@ using DynamicData.Aggregation;
 
 using DynamicTreeDataGrid.Columns;
 using DynamicTreeDataGrid.Models.Columns;
-using DynamicTreeDataGrid.State;
 
 namespace DynamicTreeDataGrid;
 
@@ -47,10 +46,6 @@ public class DynamicHierarchicalTreeDataGridSource<TModel, TModelKey> : Hierarch
 
     public IObservable<int> FilteredCount { get; }
     public IObservable<int> TotalCount { get; }
-
-    public IEnumerable<ColumnState> GetColumnStates() => throw new NotImplementedException();
-    public bool ApplyColumnStates(IEnumerable<ColumnState> states) => throw new NotImplementedException();
-
 
     public new DynamicColumnList<TModel> Columns { get; } = [];
     IDynamicColumns IDynamicTreeDataGridSource.Columns => Columns;
